product item view

+++ routes/types.js
@@ -6,6 +6,7 @@ const { Category } = require('../models/category');
 const router = express.Router();
 const auth = require('../middleware/auth');
 const admin = require('../middleware/admin');
+const { Product } = require('../models/product');
 
 router.get('/', async(req, res) => {
     const types = await Type.find().sort('name');
@@ -90,7 +91,13 @@ router.put('/:id', validateObjectId, async (req, res) => {
 });
 
 router.delete('/:id', [auth, admin], validateObjectId, async (req, res) => {
-   const type = await Type.findByIdAndRemove(req.params.id);
+    const product = await Product.find({'type._id': req.params.id});
+
+    if(product[0]) {
+        return res.status(404).send({message: 'Cannot delete this type. Products are registered in this type.'});
+    }
+
+    const type = await Type.findByIdAndRemove(req.params.id);
 
    if(!type) {
        return res.status(404).send('The category with this id is not found');
